Add more validation for registration and request data
Major refactoring

Conflicts:
	app.py

+++ dbhelper.py
@@ -1,27 +1,45 @@
 import psycopg2, psycopg2.extras
-import sys
 from flask import flash, session
 from constants import *
 import validate
 import uuid
 
+def connect():
+    try:
+        return psycopg2.connect("dbname='%s' user='%s'" % (DB_NAME, DB_USER))
+    except psycopg2.DatabaseError, e:
+        print "Error: %s" % e
+        return None
+
 def auth(type, email, password, phone=None):
     global ID_USER
-    conn = None
+    conn = connect()
+    if conn == None:
+        return "Database Error"
+    c = conn.cursor()
     try:
-        conn = psycopg2.connect("dbname='%s' user='%s'" % (DB_NAME, DB_USER))
-        c = conn.cursor()
-        if type=="register":
-            c.execute("SELECT 1 FROM Users WHERE Email = %s", (email,))
-            if c.fetchall() == []:
+        if type == AUTH_REGISTER:
+            valid_email = validate.is_valid_email(email, check_db=False)
+            if not valid_email[0]:
+                return valid_email[1]
+            valid_password = validate.is_valid_password(password)
+            if not valid_password[0]:
+                return valid_password[1]
+            valid_phone = validate.is_valid_telephone(phone)
+            if not valid_phone[0]:
+                return valid_phone[1]
+            if not email_exists(email):
+                uuid = generate_id(ID_USER)
+                if not uuid[0]:
+                    return "UUID error"
                 c.execute("INSERT INTO Users VALUES(%s, %s, %s, %s)",
-                          (generate_id(ID_USER), email, validate.generate_password_hash(password),
+                          (uuid[1], email, validate.generate_password_hash(password),
                            phone))
                 conn.commit()
                 return "Registration successful"
             else:
-                return "Username is taken"
-        elif type=="login":
+                return "A user with that email already exists"
+        elif type == AUTH_LOGIN:
             print c.execute("""SELECT Password FROM Users WHERE Email = %s""",
                             (email,))
             results = c.fetchall()
@@ -41,16 +59,20 @@ def auth(type, email, password, phone=None):
 
 def add_place(name, location_x, location_y, finder):
     global ID_PLACE
-    conn = None
+    conn = connect()
+    if conn == None:
+        return "Database Error"
+    c = conn.cursor()
     try:
-        conn = psycopg2.connect("dbname='%s' user='%s'" % (DB_NAME, DB_USER))
-        c = conn.cursor()
         #remember to change the first 0 into a random int
         c.execute("SELECT 1 FROM Places WHERE Name=%s AND LocationX=%s AND LocationY=%s", (name, location_x, location_y))
         exists = c.fetchall()
         if exists == []:
+            uuid = generate_id(ID_PLACE)
+            if not uuid[0]:
+                return "UUID error"
             c.execute("INSERT INTO Places VALUES(%s, %s, %s, %s, 0, %s)",
-                      (generate_id(ID_PLACE), name, location_x, location_y, finder))
+                      (uuid[1], name, location_x, location_y, finder))
             conn.commit()
             return "Location added to map"
         else:
@@ -62,10 +84,11 @@ def add_place(name, location_x, location_y, finder):
             conn.close()
 
 def remove_place(name, location_x, location_y):
-    conn = None
+    conn = connect()
+    if conn == None:
+        return "Database Error"
+    c = conn.cursor()
     try:
-        conn = psycopg2.connect("dbname='%s' user='%s'" % (DB_NAME, DB_USER))
-        c = conn.cursor()
         c.execute("""DELETE FROM Places WHERE Name = %s AND LocationX = %s AND
                   LocationY = %s""", (name, location_x, location_y))
         conn.commit()
@@ -77,10 +100,11 @@ def remove_place(name, location_x, location_y):
             conn.close()
 
 def get_places():
-    conn = None
+    conn = connect()
+    if conn == None:
+        return "Database Error"
+    c = conn.cursor()
     try:
-        conn = psycopg2.connect("dbname='%s' user='%s'" % (DB_NAME, DB_USER))
-        c = conn.cursor()
         c.execute("SELECT * FROM PLACES")
         conn.commit()
         return dictionarify(c.fetchall())
@@ -103,10 +127,11 @@ def dictionarify(places_list):
     return ans
 
 def get_local_places(location_x, location_y, radius):
-    conn = None
+    conn = connect()
+    if conn == None:
+        return "Database Error"
+    c = conn.cursor()
     try:
-        conn = psycopg2.connect("dbname='%s' user='%s'" % (DB_NAME, DB_USER))
-        c = conn.cursor()
         c.execute("""SELECT * FROM PLACES WHERE abs(LocationX-%s) <= %s AND
         abs(LocationY-%s) <= %s""", (location_x, radius, location_y, radius))
         #c.execute("""SELECT * FROM PLACES WHERE abs(LocationX-%s) <= 1 AND abs(LocationY-%s) <= 1""", ('3', '3'))
@@ -119,10 +144,11 @@ def get_local_places(location_x, location_y, radius):
             conn.close()
 
 def email_exists(email):
-    conn = None
+    conn = connect()
+    if conn == None:
+        return "Database Error"
+    c = conn.cursor()
     try:
-        conn = psycopg2.connect("dbname='%s' user='%s'" % (DB_NAME, DB_USER))
-        c = conn.cursor()
         c.execute("""SELECT 1 FROM Users WHERE Email = %s""", (email,))
         conn.commit()
         return len(c.fetchall()) > 0
@@ -144,19 +170,21 @@ def generate_id(datatype):
     success = False
     psycopg2.extras.register_uuid()
     u = uuid.uuid4()
+    conn = connect()
+    if conn == None:
+        return (False, "Database Error")
+    c = conn.cursor()
     try:
-        conn = psycopg2.connect("dbname='%s' user='%s'" % (DB_NAME, DB_USER))
-        c = conn.cursor()
         while not success:
             c.execute(query, (u,))
             conn.commit()
             if len(c.fetchall()) == 0:
                 success = True
-	    else:
-	        u = uuid.uuid4()
+            else:
+                u = uuid.uuid4()
     except psycopg2.DatabaseError, e:
         print 'Error %s' % e
     finally:
         if conn:
             conn.close()
-    return u
+    return (True, u)
